backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
psycopg2-binary>=2.9,<3.0
gunicorn>=21.2,<22.0
Pillow>=10.0,<11.0
django-cors-headers>=4.3,<5.0
dj-database-url>=2.1,<3.0
